move terrain to new terrain subpackage
